[C8] Split constraints_2d.go by constraint family (#2241)
model/sketch/constraints_2d.go was 613 LOC, over the 500-line ceiling. Split
the 2D geometric-constraint definitions by family; each file is now under the
ceiling:

- constraints_2d.go (268): the point/incidence constraints (coincident,
  point-on-line, midpoint, point-on-circle, horizontal, vertical, symmetry,
  fix) plus the shared residual/Jacobian AD scaffolding (lineVars, adLineDirs,
  adTwoLines).
- constraints_2d_lines.go (123): the two-line relations (parallel,
  perpendicular, collinear, equal-length).
- constraints_2d_circles.go (275): the circle relations and tangents
  (concentric, equal-radius, line-tangent, circle-tangent) and their residual
  helpers.

Behaviour-preserving: split programmatically by receiver-qualified
declaration, verified every non-comment code line is byte-identical (only
per-file import redistribution + gofmt one-liner alignment differ, the latter
restored); imports resolved per file with goimports. The file is entirely
geometric constraints (the dimensional constraints already live elsewhere), so
the split is by geometric sub-family. Sketch package builds, archguard passes,
lint 0. Net delta of recognizers / tolerance constants / fallback sites / type
assertions: 0.

Part of #2183 (C8: structure/size limits). Closes #2241.

(cherry picked from commit 0843e25d56af5f0227a71f51d700113ab037c80f)
